Decision tree voting scheme. Implements a feature selection approach based on Decision Trees, using a voting scheme across the top levels on trees trained on multiple subsamples.

Decision tree voting scheme. Implements a feature selection approach based on Decision Trees, using a voting scheme across the top levels on trees trained on multiple subsamples.

dtreevoting(num_runs = 100, num_levels = 10, file_path = file_path)

Arguments

num_runs
Number of subsamples to use for voting scheme (default: 100)

num_levels
Number of levels in each tree to consider. Only the features which appear in the top num_levels levels of the trees (from the root) will be counted

file_path
Where the num_runs subsample files are found (e.g. if sample 10 is at 'subsamples/sample10.csv' then file_path should be 'subsamples/sample'). There must be enough samples to fulfill num_runs runs.

Value

Outputs a dataframe containing (first column) total number of appearances of each feature (each row is a feature). The rest of the columns represent 1 run each and contain the level at which the feature appears.

Examples

dtreevoting(
  num_runs=5,
  num_levels=10,
  file_path=paste(system.file('samples/subsamples', package = "feamiR"),'/sample',sep=''))

#>                X1 X2 X3 X4 X5 X6
#> classification  0 NA NA NA NA NA
#> position1_A     1  3 NA NA NA NA
#> position1_C     0 NA NA NA NA NA
#> position1_G     0 NA NA NA NA NA
#> position1_U     4  3  4 NA  3  3
#> position2_A     2 NA NA NA  2  2
#> position2_C     0 NA NA NA NA NA
#> position2_G     1 NA  4 NA NA NA
#> position2_U     1 NA NA NA NA  5
#> position3_A     1 NA  5 NA NA NA
#> position3_C     0 NA NA NA NA NA
#> position3_G     0 NA NA NA NA NA
#> position3_U     1 NA NA  5 NA NA
#> position4_A     1 NA NA NA  6 NA
#> position4_C     1  7 NA NA NA NA
#> position4_G     0 NA NA NA NA NA
#> position4_U     1 NA NA  7 NA NA
#> position5_A     3  4 NA NA  3  3
#> position5_C     0 NA NA NA NA NA
#> position5_G     1 NA  3 NA NA NA
#> position5_U     2 NA  2  2 NA NA
#> position6_A     1  3 NA NA NA NA
#> position6_C     0 NA NA NA NA NA
#> position6_G     1 NA  6 NA NA NA
#> position6_U     1  2 NA NA NA NA
#> position7_A     0 NA NA NA NA NA
#> position7_C     0 NA NA NA NA NA
#> position7_G     0 NA NA NA NA NA
#> position7_U     0 NA NA NA NA NA
#> position8_A     0 NA NA NA NA NA
#> position8_C     1 NA NA NA  5 NA
#> position8_G     3  8  7 NA  5 NA
#> position8_U     0 NA NA NA NA NA
#> pair1_AA        0 NA NA NA NA NA
#> pair1_AC        0 NA NA NA NA NA
#> pair1_AG        0 NA NA NA NA NA
#> pair1_AU        2  6 NA  6 NA NA
#> pair1_CA        0 NA NA NA NA NA
#> pair1_CC        0 NA NA NA NA NA
#> pair1_CG        0 NA NA NA NA NA
#> pair1_CU        0 NA NA NA NA NA
#> pair1_GA        0 NA NA NA NA NA
#> pair1_GC        0 NA NA NA NA NA
#> pair1_GG        0 NA NA NA NA NA
#> pair1_GU        0 NA NA NA NA NA
#> pair1_UA        1 NA NA  4 NA NA
#> pair1_UC        0 NA NA NA NA NA
#> pair1_UG        1 NA NA  4 NA NA
#> pair1_UU        0 NA NA NA NA NA
#> pair2_AA        0 NA NA NA NA NA
#> pair2_AC        0 NA NA NA NA NA
#> pair2_AG        0 NA NA NA NA NA
#> pair2_AU        0 NA NA NA NA NA
#> pair2_CA        0 NA NA NA NA NA
#> pair2_CC        0 NA NA NA NA NA
#> pair2_CG        1 NA NA NA  8 NA
#> pair2_CU        0 NA NA NA NA NA
#> pair2_GA        3  5  6 NA  4 NA
#> pair2_GC        0 NA NA NA NA NA
#> pair2_GG        0 NA NA NA NA NA
#> pair2_GU        1 NA NA  3 NA NA
#> pair2_UA        0 NA NA NA NA NA
#> pair2_UC        0 NA NA NA NA NA
#> pair2_UG        0 NA NA NA NA NA
#> pair2_UU        1 NA NA NA  7 NA
#> pair3_AA        0 NA NA NA NA NA
#> pair3_AC        3  2  2 NA NA  2
#> pair3_AG        2 NA NA  5 NA  4
#> pair3_AU        0 NA NA NA NA NA
#> pair3_CA        0 NA NA NA NA NA
#> pair3_CC        0 NA NA NA NA NA
#> pair3_CG        0 NA NA NA NA NA
#> pair3_CU        0 NA NA NA NA NA
#> pair3_GA        2 NA NA NA  4  4
#> pair3_GC        0 NA NA NA NA NA
#> pair3_GG        0 NA NA NA NA NA
#> pair3_GU        0 NA NA NA NA NA
#> pair3_UA        0 NA NA NA NA NA
#> pair3_UC        0 NA NA NA NA NA
#> pair3_UG        0 NA NA NA NA NA
#> pair3_UU        1 NA  6 NA NA NA
#> pair4_AA        1 NA  4 NA NA NA
#> pair4_AC        0 NA NA NA NA NA
#> pair4_AG        0 NA NA NA NA NA
#> pair4_AU        0 NA NA NA NA NA
#> pair4_CA        1 NA  5 NA NA NA
#> pair4_CC        0 NA NA NA NA NA
#> pair4_CG        2 NA NA  2  2 NA
#> pair4_CU        0 NA NA NA NA NA
#> pair4_GA        0 NA NA NA NA NA
#> pair4_GC        0 NA NA NA NA NA
#> pair4_GG        0 NA NA NA NA NA
#> pair4_GU        1 NA NA  6 NA NA
#> pair4_UA        0 NA NA NA NA NA
#> pair4_UC        0 NA NA NA NA NA
#> pair4_UG        0 NA NA NA NA NA
#> pair4_UU        0 NA NA NA NA NA
#> pair5_AA        0 NA NA NA NA NA
#> pair5_AC        0 NA NA NA NA NA
#> pair5_AG        1 NA NA  5 NA NA
#> pair5_AU        0 NA NA NA NA NA
#> pair5_CA        0 NA NA NA NA NA
#> pair5_CC        0 NA NA NA NA NA
#> pair5_CG        0 NA NA NA NA NA
#> pair5_CU        0 NA NA NA NA NA
#> pair5_GA        0 NA NA NA NA NA
#> pair5_GC        2 NA NA NA  1  1
#> pair5_GG        0 NA NA NA NA NA
#> pair5_GU        0 NA NA NA NA NA
#> pair5_UA        0 NA NA NA NA NA
#> pair5_UC        0 NA NA NA NA NA
#> pair5_UG        0 NA NA NA NA NA
#> pair5_UU        0 NA NA NA NA NA
#> pair6_AA        0 NA NA NA NA NA
#> pair6_AC        0 NA NA NA NA NA
#> pair6_AG        0 NA NA NA NA NA
#> pair6_AU        0 NA NA NA NA NA
#> pair6_CA        0 NA NA NA NA NA
#> pair6_CC        0 NA NA NA NA NA
#> pair6_CG        0 NA NA NA NA NA
#> pair6_CU        3  3  3  3 NA NA
#> pair6_GA        0 NA NA NA NA NA
#> pair6_GC        3 NA  6 NA  6  6
#> pair6_GG        0 NA NA NA NA NA
#> pair6_GU        0 NA NA NA NA NA
#> pair6_UA        0 NA NA NA NA NA
#> pair6_UC        0 NA NA NA NA NA
#> pair6_UG        0 NA NA NA NA NA
#> pair6_UU        0 NA NA NA NA NA
#> pair7_AA        0 NA NA NA NA NA
#> pair7_AC        3  1  1  1 NA NA
#> pair7_AG        1  6 NA NA NA NA
#> pair7_AU        0 NA NA NA NA NA
#> pair7_CA        0 NA NA NA NA NA
#> pair7_CC        0 NA NA NA NA NA
#> pair7_CG        0 NA NA NA NA NA
#> pair7_CU        0 NA NA NA NA NA
#> pair7_GA        0 NA NA NA NA NA
#> pair7_GC        0 NA NA NA NA NA
#> pair7_GG        0 NA NA NA NA NA
#> pair7_GU        2  5  5 NA NA NA
#> pair7_UA        0 NA NA NA NA NA
#> pair7_UC        0 NA NA NA NA NA
#> pair7_UG        1 NA NA NA NA  5
#> pair7_UU        0 NA NA NA NA NA